Added item blocks (can be currently dropped with q)
    added water movements/propagation.
    Added player skins. Can also toggle 3rd person camera with F5
    Added a way to send multiple packets in one go. (GROUP)

Lots of refactor and more minor "improvements".

+++ include/shared/Config.hpp
@@ -10,10 +10,10 @@ constexpr float TPS = 20.0f;
 
 using TickDuration = std::chrono::steady_clock::duration;
 constexpr TickDuration TICK_RATE = std::chrono::duration_cast<TickDuration>(
-    std::chrono::duration<double>(1.0 / TPS)
+    std::chrono::duration<float>(1.0 / TPS)
 );
 
 // convert to milliseconds:
-constexpr double MS_TICK_RATE = std::chrono::duration_cast<std::chrono::milliseconds>(TICK_RATE).count();
+constexpr float MS_TICK_RATE = std::chrono::duration_cast<std::chrono::milliseconds>(TICK_RATE).count();
 
 #endif
